@@ -1,8 +1,9 @@
-export interface ShoppingListProductNg {
-    id: string;
-    amount: number;
-    unit: string;
-    productTitle: string;
-    productId: string;
-    shoppingListId: string;
+export class ShoppingListProductNg {
+    id: string = '';
+    amount: number = 0;
+    unit: string = '';
+    productTitle: string = '';
+    productId: string = '';
+    shoppingListId: string = '';
+    active:boolean = true;
 }
